Add gem 'rails-controller-testing'
This is so I can address the error below when creating controller specs
that assert that we are rendering a template for an action.

```

NoMethodError:
       assert_template has been extracted to a gem. To continue using it,
               add `gem 'rails-controller-testing'` to your Gemfile.

```

I tried adding only to test and development group, but it won't work and
I don't have much time to understand why, so for now, it's good to go.