Introduce carousel item partial
I've opted to extract this into its own partial as it's bein rendered
twice on the page. It's easier to scale in case new slides are added.

I haven't added any tests as this view does not contain any logic, it's
a purely presentational view that receives the relevant content to
render.