Add conditionals to ReportsController
I want to ensure that users have a line manager assigned to them before
they are able to create reports. An Org Admin should be able to assign
line manager to a user that belongs to their organisation.

I'm also adding a helper method that checks that a user receives a flash
message should they try to get to reports/new and attempt to create a
weekly report that was already submitted.

They can still update reports if created_at for that report still falls
on the week they are trying to update the report for.